refactor: TickerAllResponseDTO 개선 #2
