+++ Huhoe/Huhoe/Data/Network/DataMapping/TickerAllResponseDTO+Mapping.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-// MARK: - Data Transfer Object
+// MARK: - Ticker All Response DTO
 
 struct TickerAllResponseDTO: Decodable {
     let status: String
@@ -18,6 +18,42 @@ struct TickerAllResponseDTO: Decodable {
         case tickerDTO = "data"
     }
 }
+
+// MARK: - Ticker DTO
+
+extension TickerAllResponseDTO {
+    struct TickerDTO: Decodable {
+        let openingPrice: String
+        let closingPrice: String
+        let minPrice: String
+        let maxPrice: String
+        let unitsTraded: String
+        let accTradeValue: String
+        let prevClosingPrice: String
+        let unitsTraded24Hour: String
+        let accTradeValue24Hour: String
+        let fluctate24Hour: String
+        let fluctateRate24Hour: String
+        let date: String?
+        
+        enum CodingKeys: String, CodingKey {
+            case openingPrice
+            case closingPrice
+            case minPrice
+            case maxPrice
+            case unitsTraded
+            case accTradeValue
+            case prevClosingPrice
+            case date
+            case unitsTraded24Hour = "unitsTraded24H"
+            case accTradeValue24Hour = "accTradeValue24H"
+            case fluctate24Hour = "fluctate24H"
+            case fluctateRate24Hour = "fluctateRate24H"
+        }
+    }
+}
+
+// MARK: - Dynamic Value
 
 extension TickerAllResponseDTO {
     enum DynamicValue: Decodable {
